pyproject.toml
[tool.poetry]
name = "aries-vcr-vc-service"
version = "0.1.0"
description = ""
authors = ["Akiff Manji <akiff.manji@quartech.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.0"}
pytest = "^8.3.2"
httpx = "^0.27.0"
pydantic-settings = "^2.5.0"
canonicaljson = "^2.0.0"
aries-askar = "^0.3.2"
requests = "^2.32.3"
multiformats = "^0.3.1.post4"
uvicorn = "^0.30.6"
ruff = "^0.6.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
